requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
anthropic>=0.25.0
requests>=2.31.0
plotly>=5.20.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
